Create bitwarden-vault
