+++ .github/workflows/publish-rust-crates.yml
@@ -54,6 +54,11 @@ on:
         required: true
         default: true
         type: boolean
+      publish_bitwarden-vault:
+        description: "Publish bitwarden-valt crate"
+        required: true
+        default: true
+        type: boolean
 
 defaults:
   run:
@@ -91,6 +96,7 @@ jobs:
           PUBLISH_BITWARDEN_CLI: ${{ github.event.inputs.publish_bitwarden-cli }}
           PUBLISH_BITWARDEN_GENERATORS: ${{ github.event.inputs.publish_bitwarden-generators }}
           PUBLISH_BITWARDEN_EXPORTERS: ${{ github.event.inputs.publish_bitwarden-exporters }}
+          PUBLISH_BITWARDEN_VAULT: ${{ github.event.inputs.publish_bitwarden-vault }}
         run: |
           if [[ "$PUBLISH_BITWARDEN" == "false" ]] && [[ "$PUBLISH_BITWARDEN_API_API" == "false" ]] && [[ "$PUBLISH_BITWARDEN_API_IDENTITY" == "false" ]]; then
             echo "==================================="
@@ -140,6 +146,11 @@ jobs:
           if [[ "$PUBLISH_BITWARDEN_EXPORTERS" == "true" ]]; then
             PACKAGES_COMMAND="$PACKAGES_COMMAND -p bitwarden-exporters"
             PACKAGES_LIST="$PACKAGES_LIST bitwarden-exporters"
+          fi
+
+          if [[ "$PUBLISH_BITWARDEN_VAULT" == "true" ]]; then
+            PACKAGES_COMMAND="$PACKAGES_COMMAND -p bitwarden-vault"
+            PACKAGES_LIST="$PACKAGES_LIST bitwarden-vault"
           fi
 
           echo "Packages command: " $PACKAGES_COMMAND
