+++ crates/bitwarden-vault/Cargo.toml
@@ -1,0 +1,47 @@
+[package]
+name = "bitwarden-vault"
+description = """
+Internal crate for the bitwarden crate. Do not use.
+"""
+
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license-file.workspace = true
+keywords.workspace = true
+
+[features]
+uniffi = ["dep:uniffi"] # Uniffi bindings
+
+[dependencies]
+base64 = ">=0.21.2, <0.23"
+bitwarden-api-api = { workspace = true }
+bitwarden-core = { workspace = true }
+bitwarden-crypto = { workspace = true }
+bitwarden-exporters = { workspace = true }
+chrono = { version = ">=0.4.26, <0.5", features = [
+    "clock",
+    "serde",
+    "std",
+], default-features = false }
+rand = ">=0.8.5, <0.9"
+hmac = ">=0.12.1, <0.13"
+reqwest = { version = ">=0.12, <0.13", features = [
+    "http2",
+    "json",
+], default-features = false }
+schemars = { version = ">=0.8.9, <0.9", features = ["uuid1", "chrono"] }
+serde = { version = ">=1.0, <2.0", features = ["derive"] }
+serde_json = ">=1.0.96, <2.0"
+serde_repr = ">=0.1.12, <0.2"
+sha1 = ">=0.10.5, <0.11"
+sha2 = ">=0.10.6, <0.11"
+thiserror = ">=1.0.40, <2.0"
+uniffi = { version = "=0.27.2", optional = true }
+uuid = { version = ">=1.3.3, <2.0", features = ["serde"] }
+
+[lints]
+workspace = true
