+++ crates/bitwarden-vault/src/cipher/field.rs
@@ -8,7 +8,7 @@ use serde::{Deserialize, Serialize};
 use serde_repr::{Deserialize_repr, Serialize_repr};
 
 use super::linked_id::LinkedIdType;
-use crate::error::{Error, Result};
+use crate::VaultParseError;
 
 #[derive(Clone, Copy, Serialize_repr, Deserialize_repr, Debug, JsonSchema)]
 #[repr(u8)]
@@ -65,9 +65,9 @@ impl KeyDecryptable<SymmetricCryptoKey, FieldView> for Field {
 }
 
 impl TryFrom<CipherFieldModel> for Field {
-    type Error = Error;
+    type Error = VaultParseError;
 
-    fn try_from(model: CipherFieldModel) -> Result<Self> {
+    fn try_from(model: CipherFieldModel) -> Result<Self, Self::Error> {
         Ok(Self {
             name: EncString::try_from_optional(model.name)?,
             value: EncString::try_from_optional(model.value)?,
