+++ crates/bitwarden-vault/src/cipher/local_data.rs
