+++ crates/bitwarden-vault/src/cipher/mod.rs
@@ -12,8 +12,10 @@ pub(crate) mod secure_note;
 pub use attachment::{
     Attachment, AttachmentEncryptResult, AttachmentFile, AttachmentFileView, AttachmentView,
 };
-pub use cipher::{Cipher, CipherListView, CipherRepromptType, CipherType, CipherView};
+pub use cipher::{Cipher, CipherError, CipherListView, CipherRepromptType, CipherType, CipherView};
 pub use field::FieldView;
-pub(crate) use login::Fido2CredentialFullView;
-pub use login::{Fido2Credential, Fido2CredentialNewView, Fido2CredentialView};
+pub use login::{
+    Fido2Credential, Fido2CredentialFullView, Fido2CredentialNewView, Fido2CredentialView, Login,
+    LoginView,
+};
 pub use secure_note::SecureNoteType;
