+++ crates/bitwarden-vault/src/domain.rs
@@ -2,7 +2,7 @@ use bitwarden_core::require;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use crate::error::{Error, Result};
+use crate::VaultParseError;
 
 #[derive(Serialize, Deserialize, Debug, JsonSchema)]
 pub struct GlobalDomains {
@@ -12,9 +12,11 @@ pub struct GlobalDomains {
 }
 
 impl TryFrom<bitwarden_api_api::models::GlobalDomains> for GlobalDomains {
-    type Error = Error;
+    type Error = VaultParseError;
 
-    fn try_from(global_domains: bitwarden_api_api::models::GlobalDomains) -> Result<Self> {
+    fn try_from(
+        global_domains: bitwarden_api_api::models::GlobalDomains,
+    ) -> Result<Self, Self::Error> {
         Ok(Self {
             r#type: require!(global_domains.r#type),
             domains: require!(global_domains.domains),
