+++ crates/bitwarden-vault/src/folder.rs
@@ -8,7 +8,7 @@ use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
-use crate::error::{Error, Result};
+use crate::VaultParseError;
 
 #[derive(Serialize, Deserialize, Debug, JsonSchema)]
 #[serde(rename_all = "camelCase")]
@@ -49,9 +49,9 @@ impl KeyDecryptable<SymmetricCryptoKey, FolderView> for Folder {
 }
 
 impl TryFrom<FolderResponseModel> for Folder {
-    type Error = Error;
+    type Error = VaultParseError;
 
-    fn try_from(folder: FolderResponseModel) -> Result<Self> {
+    fn try_from(folder: FolderResponseModel) -> Result<Self, Self::Error> {
         Ok(Folder {
             id: folder.id,
             name: require!(EncString::try_from_optional(folder.name)?),
