+++ crates/bitwarden-vault/src/uniffi_support.rs
@@ -1,0 +1,16 @@
+use std::num::NonZeroU32;
+
+use bitwarden_crypto::{AsymmetricEncString, EncString};
+use uuid::Uuid;
+
+uniffi::ffi_converter_forward!(NonZeroU32, bitwarden_crypto::UniFfiTag, crate::UniFfiTag);
+uniffi::ffi_converter_forward!(EncString, bitwarden_crypto::UniFfiTag, crate::UniFfiTag);
+uniffi::ffi_converter_forward!(
+    AsymmetricEncString,
+    bitwarden_crypto::UniFfiTag,
+    crate::UniFfiTag
+);
+
+type DateTime = chrono::DateTime<chrono::Utc>;
+uniffi::ffi_converter_forward!(DateTime, bitwarden_core::UniFfiTag, crate::UniFfiTag);
+uniffi::ffi_converter_forward!(Uuid, bitwarden_core::UniFfiTag, crate::UniFfiTag);
