+++ crates/bitwarden/src/auth/password/validate.rs
@@ -43,6 +43,8 @@ pub(crate) fn validate_password_user_key(
     password: String,
     encrypted_user_key: String,
 ) -> Result<String> {
+    use bitwarden_core::VaultLocked;
+
     let login_method = client
         .login_method
         .as_ref()
@@ -59,9 +61,9 @@ pub(crate) fn validate_password_user_key(
 
                 let enc = client
                     .get_encryption_settings()
-                    .map_err(|_| Error::VaultLocked)?;
+                    .map_err(|_| VaultLocked())?;
 
-                let existing_key = enc.get_key(&None).ok_or(Error::VaultLocked)?;
+                let existing_key = enc.get_key(&None).ok_or(VaultLocked())?;
 
                 if user_key.to_vec() != existing_key.to_vec() {
                     return Err("wrong user key".into());
