+++ crates/bitwarden/src/mobile/crypto.rs
@@ -1,5 +1,6 @@
 use std::collections::HashMap;
 
+use bitwarden_core::VaultLocked;
 use bitwarden_crypto::{AsymmetricEncString, EncString};
 #[cfg(feature = "internal")]
 use bitwarden_crypto::{KeyDecryptable, KeyEncryptable, MasterKey, SymmetricCryptoKey};
@@ -175,7 +176,7 @@ pub async fn get_user_encryption_key(client: &mut Client) -> Result<String> {
     let user_key = client
         .get_encryption_settings()?
         .get_key(&None)
-        .ok_or(Error::VaultLocked)?;
+        .ok_or(VaultLocked())?;
 
     Ok(user_key.to_base64())
 }
@@ -198,7 +199,7 @@ pub fn update_password(
     let user_key = client
         .get_encryption_settings()?
         .get_key(&None)
-        .ok_or(Error::VaultLocked)?;
+        .ok_or(VaultLocked())?;
 
     let login_method = client
         .login_method
@@ -243,7 +244,7 @@ pub fn derive_pin_key(client: &mut Client, pin: String) -> Result<DerivePinKeyRe
     let user_key = client
         .get_encryption_settings()?
         .get_key(&None)
-        .ok_or(Error::VaultLocked)?;
+        .ok_or(VaultLocked())?;
 
     let login_method = client
         .login_method
@@ -263,7 +264,7 @@ pub fn derive_pin_user_key(client: &mut Client, encrypted_pin: EncString) -> Res
     let user_key = client
         .get_encryption_settings()?
         .get_key(&None)
-        .ok_or(Error::VaultLocked)?;
+        .ok_or(VaultLocked())?;
 
     let pin: String = encrypted_pin.decrypt_with_key(user_key)?;
     let login_method = client
@@ -301,7 +302,7 @@ pub(super) fn enroll_admin_password_reset(
 
     let public_key = AsymmetricPublicCryptoKey::from_der(&STANDARD.decode(public_key)?)?;
     let enc = client.get_encryption_settings()?;
-    let key = enc.get_key(&None).ok_or(Error::VaultLocked)?;
+    let key = enc.get_key(&None).ok_or(VaultLocked())?;
 
     Ok(AsymmetricEncString::encrypt_rsa2048_oaep_sha1(
         &key.to_vec(),
