+++ crates/bitwarden/src/platform/fido2/traits.rs
@@ -1,10 +1,8 @@
+use bitwarden_vault::{Cipher, CipherView, Fido2CredentialNewView};
 use passkey::authenticator::UIHint;
 use thiserror::Error;
 
-use crate::{
-    error::Result,
-    vault::{Cipher, CipherView, Fido2CredentialNewView},
-};
+use crate::error::Result;
 
 #[derive(Debug, Error)]
 pub enum Fido2CallbackError {
