+++ crates/bitwarden/src/secrets_manager/secrets/update.rs
@@ -1,14 +1,12 @@
 use bitwarden_api_api::models::SecretUpdateRequestModel;
+use bitwarden_core::VaultLocked;
 use bitwarden_crypto::KeyEncryptable;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
 use super::SecretResponse;
-use crate::{
-    client::Client,
-    error::{Error, Result},
-};
+use crate::{client::Client, error::Result};
 
 #[derive(Serialize, Deserialize, Debug, JsonSchema)]
 #[serde(rename_all = "camelCase", deny_unknown_fields)]
@@ -31,7 +29,7 @@ pub(crate) async fn update_secret(
     let key = client
         .get_encryption_settings()?
         .get_key(&Some(input.organization_id))
-        .ok_or(Error::VaultLocked)?;
+        .ok_or(VaultLocked())?;
 
     let secret = Some(SecretUpdateRequestModel {
         key: input.key.clone().encrypt_with_key(key)?.to_string(),
