+++ crates/bitwarden/src/vault/mod.rs
@@ -1,30 +1,9 @@
-mod cipher;
-pub use cipher::*;
-
-#[cfg(feature = "internal")]
-mod client_vault;
-#[cfg(feature = "internal")]
-pub use client_vault::ClientVault;
-
-mod collection;
-pub use collection::{Collection, CollectionView};
-
-mod folder;
-pub use folder::{Folder, FolderView};
-
-mod password_history;
-pub use password_history::{PasswordHistory, PasswordHistoryView};
-
-#[cfg(feature = "internal")]
 mod sync;
-#[cfg(feature = "internal")]
 pub use sync::{SyncRequest, SyncResponse};
-#[cfg(feature = "internal")]
-mod domain;
-
-#[cfg(feature = "internal")]
-mod totp;
-#[cfg(feature = "internal")]
-pub(crate) use totp::generate_totp;
-#[cfg(feature = "internal")]
-pub use totp::TotpResponse;
+mod client_vault;
+pub use bitwarden_vault::{
+    Attachment, AttachmentEncryptResult, AttachmentView, Cipher, CipherListView, CipherView,
+    Collection, CollectionView, Fido2CredentialNewView, Fido2CredentialView, Folder, FolderView,
+    PasswordHistory, PasswordHistoryView, TotpResponse,
+};
+pub use client_vault::ClientVault;
