+++ crates/bitwarden/src/vault/sync.rs
@@ -2,16 +2,15 @@ use bitwarden_api_api::models::{
     DomainsResponseModel, ProfileOrganizationResponseModel, ProfileResponseModel, SyncResponseModel,
 };
 use bitwarden_core::require;
+use bitwarden_vault::{Cipher, Collection, Folder, GlobalDomains};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
-use super::domain::GlobalDomains;
 use crate::{
     admin_console::Policy,
     client::{encryption_settings::EncryptionSettings, Client},
     error::{Error, Result},
-    vault::{Cipher, Collection, Folder},
 };
 
 #[derive(Serialize, Deserialize, Debug, JsonSchema)]
@@ -141,14 +140,15 @@ impl ProfileResponse {
 
 impl TryFrom<DomainsResponseModel> for DomainResponse {
     type Error = Error;
-    fn try_from(value: DomainsResponseModel) -> Result<Self> {
+
+    fn try_from(value: DomainsResponseModel) -> Result<Self, Self::Error> {
         Ok(Self {
             equivalent_domains: value.equivalent_domains.unwrap_or_default(),
             global_equivalent_domains: value
                 .global_equivalent_domains
                 .unwrap_or_default()
                 .into_iter()
-                .map(|s| s.try_into())
+                .map(|s| s.try_into().map_err(Error::VaultParse))
                 .collect::<Result<Vec<GlobalDomains>>>()?,
         })
     }
